refactor: extract SidebarShortcutRow and SidebarShortcutList from sidebar
Sidebar reduced from 704 to 460 lines
